Deleted team page and added small intro it in about page
